// UI/src/app/Enums/FieldIntegration.enum.ts
export enum ATSPropertyDatatypeEnum {
    Bool = 105,
    Date = 106,
    Double = 107,
    Int = 108,
    Multiline = 109,
    MultiSelect = 110,
    RefValue = 111,
    String = 112,
    Availability = 113
}

export enum OnboardedObjects {
    Candidate = 114,
    Skills = 115,
    LicencesandCertificates = 116,
    WorkHistory = 117,
    Payroll = 118,
    Induction = 119,
    CustomFields = 120,
    Documents = 121,
    CoreFields = 139

}

export enum IntegrationType {
    CodeHouse = 'CodeHouse',
    FastTrack = 'FastTrack',
    Nero = 'Nero',
    JobAdder = 'JobAdder',
    Oncore = 'Oncore',
    StandAlone = 'StandAlone',
    Bullhorn = 'Bullhorn',
    Rdb = 'Rdb',
    Vincere = 'Vincere',
    Salesforce = 'Salesforce',
    ARCRM = 'ARCRM',
    Lumary = 'Lumary'

}